Added feedback popup to common and css changes
