+++ web/ui-app-v2/src/components/Checkbox/index.css
@@ -1,3 +1,3 @@
 .checkbox-container {
-  padding-left: 47px;
+  padding-left: 18%;
 }
